refactor: Move IndexPerformanceDto
